requirements.txt
altair==4.0.1
appnope==0.1.0
astor==0.8.1
attrs==19.3.0
backcall==0.1.0
base58==2.0.0
bleach==3.1.1
blinker==1.4
boto3==1.12.7
botocore==1.15.7
certifi==2019.11.28
chardet==3.0.4
Click==7.0
decorator==4.4.1
defusedxml==0.6.0
docutils==0.15.2
entrypoints==0.3
enum-compat==0.0.3
future==0.18.2
idna==2.9
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.12.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.16.0
Jinja2==2.11.1
jmespath==0.9.5
jsonschema==3.2.0
jupyter-client==6.0.0
jupyter-core==4.6.3
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
numpy==1.18.1
pandas==1.0.1
pandocfilters==1.4.2
parso==0.6.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
prometheus-client==0.7.1
prompt-toolkit==3.0.3
protobuf==3.11.3
ptyprocess==0.6.0
pydeck==0.2.1
Pygments==2.5.2
pyrsistent==0.15.7
python-dateutil==2.8.0
pytz==2019.3
pyzmq==19.0.0
regex==2017.4.5
requests==2.23.0
s3transfer==0.3.3
Send2Trash==1.5.0
six==1.14.0
streamlit==0.56.0
terminado==0.8.3
testpath==0.4.4
toml==0.10.0
toolz==0.10.0
torch==1.4.0
tornado==5.1.1
tqdm==4.43.0
traitlets==4.3.3
tzlocal==2.0.0
urllib3==1.25.8
validators==0.14.2
watchdog==0.10.2
wcwidth==0.1.8
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==3.0.0
